added command and layer modules
