+++ layer/__init__.py
@@ -1,0 +1,3 @@
+from .BallLayer import BallLayer
+from .PaddleLayer import PaddleLayer
+from .TileLayer import TileLayer
