+++ properties.py
@@ -33,3 +33,9 @@ CAVITY = "CAVITY"
 PROGRESSIVE = "PROGRESSIVE"
 
 SCOREBOARD = "SCOREBOARD"
+
+# UNIT STATUS
+
+UNIT_STATUS_ALIVE = "ALIVE"
+
+UNIT_STATUS_DESTROYED = "DESTROYED"
