+++ state/GameItem.py
@@ -1,34 +1,13 @@
 from pygame.draw import rect
 from pygame.math import Vector2
 from pygame import Rect
+from properties import UNIT_STATUS_ALIVE
 
 
 class GameItem(Rect):
     def __init__(self, state, position):
         self.state = state
-        self.status = "alive"
+        self.status = UNIT_STATUS_ALIVE
         self.position = position
-        self.type = None
         self.height = 0
         self.width = 0
-        self.shape = Rect(self.position, (self.width, self.height))
-
-    def move(self, moveVector):
-        raise NotImplementedError
-
-    def render(self, color):
-        rect(self.window, color, self.shape)
-
-    def getXBounds(self):
-        x = self.position.x
-        return x, x + self.width
-
-    def getYBounds(self):
-        y = self.position.y
-        return y, y + self.height
-
-    def getCenter(self):
-        return Vector2(
-            int(self.position.x + self.width / 2),
-            int(self.position.y + self.height / 2),
-        )
